Add description, filter before iteration for loop and take() in cheatcode context creation
